LayoutTests/inspector/css/stylesheet-events-inspector-stylesheet.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("CSS.StyleSheetEvents.InspectorStyleSheet");

    let bodyNodeId;

    suite.addTestCase({
        name: "CheckNoStyleSheets",
        description: "Ensure there are currently no stylesheets.",
        test: (resolve, reject) => {
            InspectorTest.expectThat(WebInspector.cssStyleManager.styleSheets.length === 0, "Should be no stylesheets.");
            resolve();
        }
    });

    suite.addTestCase({
        name: "CreateInspectorStyleSheet",
        description: "Creating an inspector stylesheet adds a stylesheet.",
        test: (resolve, reject) => {
            WebInspector.cssStyleManager.singleFireEventListener(WebInspector.CSSStyleManager.Event.StyleSheetAdded, function(event) {
                InspectorTest.expectThat(WebInspector.cssStyleManager.styleSheets.length === 1, "Should be one stylesheet.");
                InspectorTest.assert(event.data.styleSheet instanceof WebInspector.CSSStyleSheet, "Event data should be a CSSStyleSheet");
                InspectorTest.expectThat(event.data.styleSheet.origin === WebInspector.CSSStyleSheet.Type.Inspector, "StyleSheet origin is inspector.");
                resolve();
            });

            // FIXME: Currently the only way to create an inspector stylesheet is through `CSS.addRule`.
            CSSAgent.addRule(bodyNodeId, "body");
        }
    });

    WebInspector.domTreeManager.requestDocument(function(documentNode) {
        WebInspector.domTreeManager.querySelector(documentNode.id, "body", function(contentNodeId) {
            bodyNodeId = contentNodeId;
            suite.runTestCasesAndFinish();
        });
    });
}
</script>
</head>
<body onload="runTest()">
    <p>Test for CSS.styleSheetAdded for Inspector stylesheets.</p>
</body>
</html>
